Refactor JUnit 5 tests consolidate BaseTestAESGCMCICOWithGCM
This change consolidates various tests associated with the former BaseTestAESGCMCICOWithGCM into a single parameterized test.

Fixes: https://github.ibm.com/runtimes/jit-crypto/issues/1074
